Display each test case individually in test runner output.
