+++ typescript/tests/TennisTest.test.ts
@@ -1,23 +1,58 @@
-import { TennisGame1, TennisGame2, TennisGame3 } from '../src';
-import { checkAllScores } from './utils';
+import * as fs from 'fs';
+import * as path from 'path';
+import { expect } from 'chai';
+import { TennisGame, TennisGame1, TennisGame2, TennisGame3 } from '../src';
 
+function getAllScores(): Array<[number, number, string]> {
+  const testCases = path.resolve(__dirname, 'scores.json');
+  const scoreData = fs.readFileSync(testCases).toString();
+  try {
+    const scores = JSON.parse(scoreData);
+    return JSON.parse(JSON.stringify(scores));
+  } catch (err) {
+    throw new Error(`There was an error parsing the scores: "${err.message}"`);
+  }
+}
+
+const scores: Array<[number, number, string]> = getAllScores();
+
+function checkScore(game: TennisGame, player1Score: number, player2Score: number, expectedScore: string): void {
+  const highestScore: number = Math.max(player1Score, player2Score);
+  for (let i = 0; i < highestScore; i++) {
+    if (i < player1Score) {
+      game.wonPoint('player1');
+    }
+    if (i < player2Score) {
+      game.wonPoint('player2');
+    }
+  }
+  expect(game.getScore()).to.be.equals(expectedScore);
+}
 
 describe('TennisGame', function () {
+
   describe('TennisGame1', function () {
-    it('should correctly check all the scores for game 1', function () {
-      checkAllScores(() => new TennisGame1('player1', 'player2'));
+    scores.forEach(([player1Score, player2Score, expectedScore]) => {
+      it(`scores ${player1Score}:${player2Score} as ${expectedScore}`, function () {
+        checkScore(new TennisGame1('player1', 'player2'), player1Score, player2Score, expectedScore);
+      });
     });
   });
 
   describe('TennisGame2', function () {
-    it('should correctly check all the scores for game 2', function () {
-      checkAllScores(() => new TennisGame2('player1', 'player2'));
+    scores.forEach(([player1Score, player2Score, expectedScore]) => {
+      it(`scores ${player1Score}:${player2Score} as ${expectedScore}`, function () {
+        checkScore(new TennisGame2('player1', 'player2'), player1Score, player2Score, expectedScore);
+      });
     });
   });
 
   describe('TennisGame3', function () {
-    it('should correctly check all the scores for game 3', function () {
-      checkAllScores(() => new TennisGame3('player1', 'player2'));
+    scores.forEach(([player1Score, player2Score, expectedScore]) => {
+      it(`scores ${player1Score}:${player2Score} as ${expectedScore}`, function () {
+        checkScore(new TennisGame3('player1', 'player2'), player1Score, player2Score, expectedScore);
+      });
     });
   });
+
 });
